Seperate the navigation into components to improve performance
